move some components to components folder
