+++ src/components/CurrencyNumberInput/index.jsx
@@ -1,0 +1,2 @@
+import CurrencyNumberInput from "./CurrencyNumberInput";
+export { CurrencyNumberInput };
