+++ src/components/DropDownSelect/index.jsx
@@ -1,0 +1,2 @@
+import DropDownSelect from "./DropDownSelect";
+export { DropDownSelect };
