+++ src/components/PhoneNumberInput/index.jsx
@@ -1,0 +1,2 @@
+import PhoneNumberInput from "./PhoneNumberInput";
+export { PhoneNumberInput };
