+++ src/components/StandardTextField/index.jsx
@@ -1,0 +1,2 @@
+import StandardTextField from "./StandardTextField";
+export { StandardTextField }
